# backend/internal/dispute/infrastructure/postgres/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlcgen

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
	"github.com/shopspring/decimal"
)

type Account struct {
	ID         uuid.UUID
	HolderName string
	Currency   string
	OpenedAt   time.Time
	TenantID   uuid.UUID
}

type Dispute struct {
	ID             uuid.UUID
	Regime         string
	State          string
	Appeals        int32
	Version        int64
	TransactionID  uuid.UUID
	AccountID      uuid.UUID
	DisputedAmount decimal.Decimal
	Currency       string
	OpenedAt       time.Time
	UpdatedAt      time.Time
	TenantID       uuid.UUID
}

type DisputeEvent struct {
	ID             int64
	DisputeID      uuid.UUID
	Seq            int32
	Event          string
	FromState      string
	ToState        string
	Actor          string
	Payload        []byte
	IdempotencyKey *string
	TraceID        *string
	OccurredAt     time.Time
	TenantID       uuid.UUID
}

type DisputesByState struct {
	TenantID uuid.UUID
	Regime   string
	State    string
	N        int64
}

type IdempotencyKey struct {
	Scope       string
	Key         string
	RequestHash []byte
	StatusCode  int32
	Response    []byte
	CreatedAt   time.Time
	TenantID    uuid.UUID
}

type Tenant struct {
	ID           uuid.UUID
	Name         string
	Settings     []byte
	CreatedAt    time.Time
	Slug         string
	OidcIssuer   *string
	DisabledAt   pgtype.Timestamptz
	EmailDomains []string
}

type TenantKey struct {
	ID         uuid.UUID
	TenantID   uuid.UUID
	KeyHash    []byte
	Label      string
	CreatedAt  time.Time
	RevokedAt  pgtype.Timestamptz
	Prefix     string
	LastUsedAt pgtype.Timestamptz
	ExpiresAt  pgtype.Timestamptz
}

type TenantRateWindow struct {
	TenantID    uuid.UUID
	WindowStart time.Time
	Count       int32
}

type Transaction struct {
	ID         uuid.UUID
	AccountID  uuid.UUID
	Rail       string
	Amount     decimal.Decimal
	Currency   string
	Merchant   string
	OccurredAt time.Time
	TenantID   uuid.UUID
}

type WebSession struct {
	ID         uuid.UUID
	TenantID   pgtype.UUID
	Ciphertext []byte
	ExpiresAt  time.Time
	CreatedAt  time.Time
	UpdatedAt  time.Time
	Subject    *string
	Sid        *string
}