# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Dispute Engine API
  version: 0.1.0
  description: Issuer-side dispute lifecycle. Every mutation appends to the dispute's event log.
servers:
  - url: http://localhost:8090
# Every operation needs a tenant key unless it opts out with an empty security list.
security:
  - bearerAuth: []
paths:
  /healthz:
    get:
      operationId: getHealthz
      summary: Liveness
      tags: [system]
      security: []
      responses:
        "200":
          description: Process is up.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
  /readyz:
    get:
      operationId: getReadyz
      summary: Readiness (database reachable)
      tags: [system]
      security: []
      responses:
        "200":
          description: Ready.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
        "503":
          description: A dependency is unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
  /disputes:
    post:
      operationId: createDispute
      summary: Open a dispute against a transaction
      tags: [disputes]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDisputeRequest" }
      responses:
        "201":
          description: Dispute opened. Replays of the same Idempotency-Key return this same body.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispute" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /disputes/{disputeId}:
    get:
      operationId: getDispute
      summary: Current state and full event log
      tags: [disputes]
      parameters:
        - $ref: "#/components/parameters/DisputeId"
      responses:
        "200":
          description: The dispute.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispute" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /disputes/{disputeId}/events:
    post:
      operationId: applyDisputeEvent
      summary: Apply a lifecycle event
      tags: [disputes]
      parameters:
        - $ref: "#/components/parameters/DisputeId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplyEventRequest" }
      responses:
        "200":
          description: Event applied; the dispute after the transition.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispute" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "409":
          description: The event is not allowed from the current state, or a concurrent update won.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /internal/sessions/{sessionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      operationId: putSession
      summary: Store or replace an opaque session blob
      tags: [internal]
      security: [{ serviceKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SessionBlob" }
      responses:
        "204": { description: Stored. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    get:
      operationId: getSession
      summary: Fetch a session blob that has not expired
      tags: [internal]
      security: [{ serviceKey: [] }]
      responses:
        "200":
          description: The blob.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionBlob" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteSession
      summary: Remove a session
      tags: [internal]
      security: [{ serviceKey: [] }]
      responses:
        "204": { description: Removed, or never existed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /internal/sessions:
    delete:
      operationId: deleteSessions
      summary: End every session matching a realm session id or a tenant's subject
      tags: [internal]
      security: [{ serviceKey: [] }]
      parameters:
        - { name: sid, in: query, schema: { type: string } }
        - { name: tenantId, in: query, schema: { type: string, format: uuid } }
        - { name: subject, in: query, schema: { type: string } }
      responses:
        "200":
          description: How many sessions were ended.
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /internal/tenants:
    get:
      operationId: listTenants
      summary: Active tenants with what the sign-in page needs to find their realm
      tags: [internal]
      security: [{ serviceKey: [] }]
      responses:
        "200":
          description: Active tenants.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TenantSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A tenant key (issued per tenant, shown once, stored hashed) or an analyst access token from the tenant's realm.
    serviceKey:
      type: apiKey
      in: header
      name: X-Service-Key
      description: Shared secret for the frontend server's internal endpoints; never issued to tenants or browsers.
  parameters:
    DisputeId:
      name: disputeId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Client-chosen key. Same key and body replays the original response; same key with a different body is 422.
      schema: { type: string, minLength: 1, maxLength: 128 }
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    TooManyRequests:
      description: The tenant's request budget for the current minute is spent; retryAfterSeconds says when to try again.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: Missing, unknown or revoked tenant key.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unprocessable:
      description: Well-formed but not acceptable (idempotency key reused with a different body, or no regime for the transaction).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok, unavailable] }
        checks:
          type: object
          additionalProperties: { type: string }
    ErrorCode:
      type: string
      description: Stable machine-readable failure identifier; branch on this, never on text.
      enum:
        - unauthenticated
        - rate-limited
        - malformed-request
        - contract-violation
        - not-found
        - invalid-transition
        - appeals-exhausted
        - concurrent-update
        - idempotency-key-reuse
        - no-regime
        - unknown-regime
        - unavailable
        - internal
    Problem:
      type: object
      description: RFC 9457 problem details. Every field is safe to show to a user; diagnostics stay in server logs under requestId.
      required: [type, title, status, code, retryable, requestId]
      properties:
        type: { type: string, description: "urn:dispute-engine:error:<code>" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { $ref: "#/components/schemas/ErrorCode" }
        retryable:
          type: boolean
          description: true when the same request (unavailable) or a re-evaluated one (concurrent-update) can succeed later.
        retryAfterSeconds: { type: integer }
        requestId: { type: string }
        allowedEvents:
          type: array
          description: Present on invalid-transition; what the dispute accepts right now.
          items: { $ref: "#/components/schemas/DisputeEvent" }
        errors:
          type: array
          description: Field-level failures for contract-violation and business validation; field is a JSON pointer into the body, or query.x / header.x / path.x.
          items: { $ref: "#/components/schemas/FieldError" }
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string, example: "/transactionId" }
        message: { type: string }
    TenantSummary:
      type: object
      required: [id, slug, name, emailDomains]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        issuer: { type: string, description: OIDC issuer of the tenant's realm. }
        emailDomains: { type: array, items: { type: string }, description: Work-email domains that map to this tenant at sign-in. }
    SessionBlob:
      type: object
      description: Ciphertext the frontend server produced; the API stores it without being able to read it.
      required: [ciphertext, expiresAt]
      properties:
        tenantId: { type: string, format: uuid, description: Set once the session belongs to a signed-in analyst; informational. }
        subject: { type: string, description: The analyst's subject at the realm, so all of a user's sessions can be ended. }
        sid: { type: string, description: The realm's session id from the ID token, for back-channel logout. }
        ciphertext: { type: string, format: byte, maxLength: 16384 }
        expiresAt: { type: string, format: date-time }
    Regime:
      type: string
      enum: [EU_SEPA_DIRECT_DEBIT, EU_PSD2_CARD, US_REG_E, US_REG_Z]
    DisputeState:
      type: string
      enum:
        - INITIATED
        - INVESTIGATING
        - QUESTIONNAIRE_SENT
        - QUESTIONNAIRE_RECEIVED
        - PROVISIONAL_CREDIT_ISSUED
        - FAST_REFUND_ISSUED
        - SEPA_NQA_REFUND_ISSUED
        - CHARGEBACK_FILED
        - CHARGEBACK_ACKNOWLEDGED
        - EVIDENCE_SUBMITTED
        - CHARGEBACK_WON
        - CHARGEBACK_LOST
        - FINAL_CREDIT_ISSUED
        - PROVISIONAL_CREDIT_REVERSED
        - CLOSED
    DisputeEvent:
      type: string
      enum:
        - OPEN_INVESTIGATION
        - SEND_QUESTIONNAIRE
        - RECEIVE_QUESTIONNAIRE
        - ISSUE_REFUND
        - FILE_CHARGEBACK
        - ACKNOWLEDGE_CHARGEBACK
        - SUBMIT_EVIDENCE
        - WIN_CHARGEBACK
        - LOSE_CHARGEBACK
        - ISSUE_FINAL_CREDIT
        - REVERSE_PROVISIONAL_CREDIT
        - CLOSE
        - APPEAL
    CreateDisputeRequest:
      type: object
      required: [transactionId]
      properties:
        transactionId: { type: string, format: uuid }
        actor: { type: string, description: "Who opened it; defaults to customer.", default: customer }
    ApplyEventRequest:
      type: object
      required: [event]
      properties:
        event: { $ref: "#/components/schemas/DisputeEvent" }
        actor: { type: string, default: system }
        payload:
          type: object
          description: Event-specific facts, stored verbatim on the log entry.
          additionalProperties: true
    LoggedEvent:
      type: object
      required: [seq, event, fromState, toState, actor, payload, occurredAt]
      properties:
        seq: { type: integer }
        event: { type: string, description: "A DisputeEvent, or OPENED for entry 1." }
        fromState: { type: string }
        toState: { $ref: "#/components/schemas/DisputeState" }
        actor: { type: string }
        payload: { type: object, additionalProperties: true }
        traceId: { type: string }
        occurredAt: { type: string, format: date-time }
    Dispute:
      type: object
      required: [id, regime, state, appeals, version, transactionId, accountId, disputedAmount, currency, openedAt, updatedAt, allowedEvents, events]
      properties:
        id: { type: string, format: uuid }
        regime: { $ref: "#/components/schemas/Regime" }
        state: { $ref: "#/components/schemas/DisputeState" }
        appeals: { type: integer }
        version: { type: integer, format: int64 }
        transactionId: { type: string, format: uuid }
        accountId: { type: string, format: uuid }
        disputedAmount: { type: string, description: "Decimal as a string; never a float.", example: "125.4000" }
        currency: { type: string, minLength: 3, maxLength: 3 }
        openedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        allowedEvents:
          type: array
          items: { $ref: "#/components/schemas/DisputeEvent" }
        events:
          type: array
          items: { $ref: "#/components/schemas/LoggedEvent" }